Refactor: Change id to long, use snowflake id

+++ BackendDotNet/FF.Articles.Backend/FF.Articles.Backend.AI.API/Interfaces/Services/RemoteServices/IContentsApiRemoteService.cs
@@ -1,11 +1,11 @@
-    using FF.Articles.Backend.Common.ApiDtos;
+using FF.Articles.Backend.Common.ApiDtos;
 
 namespace FF.Articles.Backend.AI.API.Interfaces.Services.RemoteServices;
 
 public interface IContentsApiRemoteService
 {
-    Task<Dictionary<int, string>> AddBatchArticlesAsync(List<ArticleApiAddRequest> requests, HttpRequest httpRequest);
-    Task<bool> EditContentBatchAsync(Dictionary<int, string> batchEditConentRequests, HttpRequest httpRequest);
-    Task<int> AddTopicByTitleAsync(string title, HttpRequest httpRequest);
+    Task<Dictionary<long, string>> AddBatchArticlesAsync(List<ArticleApiAddRequest> requests, HttpRequest httpRequest);
+    Task<bool> EditContentBatchAsync(Dictionary<long, string> batchEditConentRequests, HttpRequest httpRequest);
+    Task<long> AddTopicByTitleAsync(string title, HttpRequest httpRequest);
     Task<bool> EditArticleByRequest(ArticleApiEditRequest request, HttpRequest httpRequest);
 }
