+++ BackendDotNet/FF.Articles.Backend/FF.Articles.Backend.AI.API/Services/Stores/UserArticleStateStore.cs
@@ -5,14 +5,14 @@ using FF.Articles.Backend.AI.API.Models.AiDtos;
 namespace FF.Articles.Backend.AI.API.Services.Stores;
 public class UserArticleStateStore
 {
-    private readonly ConcurrentDictionary<int, UserArticleState> _userStates = new();
+    private readonly ConcurrentDictionary<long, UserArticleState> _userStates = new();
 
-    public UserArticleState GetOrAddState(int userId)
+    public UserArticleState GetOrAddState(long userId)
     {
         return _userStates.GetOrAdd(userId, _ => new UserArticleState());
     }
 
-    public void RemoveState(int userId)
+    public void RemoveState(long userId)
     {
         _userStates.TryRemove(userId, out _);
     }
@@ -22,8 +22,8 @@ public class UserArticleState
 {
     public ArticlesAIResponse? ArticlesAIResponse { get; set; }
     // AI generated article id (sort number) => Db article id
-    public Dictionary<int, int> IdMap { get; set; } = new();
+    public Dictionary<int, long> IdMap { get; set; } = new();
     public string Topic { get; set; } = "";
-    public int TopicId {get; set; } = 0;
+    public long TopicId { get; set; } = 0;
     public bool IsFirstRound { get; set; } = true;
 }
