+++ BackendDotNet/FF.Articles.Backend/FF.Articles.Backend.Common/ApiDtos/ArticleApiAddRequest.cs
@@ -11,9 +11,9 @@ public class ArticleApiAddRequest
     ///  Article = 1, SubArticle = 2, TopicArticle = 3
     /// </summary>
     public string ArticleType { get; set; }
-    public int? ParentArticleId { get; set; }
-    public int TopicId { get; set; }
-    public List<int> TagIds { get; set; } = new();
+    public long? ParentArticleId { get; set; }
+    public long TopicId { get; set; }
+    public List<long> TagIds { get; set; } = new();
     public int SortNumber { get; set; } = 1;
     public int IsHidden { get; set; } = 0;
 }
