+++ BackendDotNet/FF.Articles.Backend/FF.Articles.Backend.Common/ApiDtos/ArticleApiEditRequest.cs
@@ -4,7 +4,7 @@ namespace FF.Articles.Backend.Common.ApiDtos;
 
 public class ArticleApiEditRequest
 {
-    public int ArticleId { get; set; }
+    public long ArticleId { get; set; }
     public string? Title { get; set; }
     public string? Content { get; set; }
     public string? Abstract { get; set; }
@@ -12,9 +12,9 @@ public class ArticleApiEditRequest
     ///  Article = 1, SubArticle = 2, TopicArticle = 3
     /// </summary>
     public string? ArticleType { get; set; }
-    public int? ParentArticleId { get; set; }
-    public int? TopicId { get; set; }
-    public List<int>? TagIds { get; set; }
+    public long? ParentArticleId { get; set; }
+    public long? TopicId { get; set; }
+    public List<long>? TagIds { get; set; }
     public int? SortNumber { get; set; }
     public int? IsHidden { get; set; }
 }
