+++ BackendDotNet/FF.Articles.Backend/FF.Articles.Backend.Contents.API/Interfaces/Repositories/V1/IArticleTagRepository.cs
@@ -5,14 +5,14 @@ using FF.Articles.Backend.Contents.API.Models.Entities;
 namespace FF.Articles.Backend.Contents.API.Interfaces.Repositories.V1;
 public interface IArticleTagRepository : IBaseRepository<ArticleTag, ContentsDbContext>
 {
-    Task<bool> EditArticleTags(int articleId, List<int> tagIds);
-    Task<bool> DeleteByArticleId(int articleId);
-    Task<bool> DeleteByTagId(int tagId);
-    Task<List<ArticleTag>> GetByArticleId(int articleId);
-    Task<List<ArticleTag>> GetByTagId(int tagId);
-    Task<List<ArticleTag>> GetByArticleIds(List<int> articleIds);
-    Task<List<ArticleTag>> GetByTagIds(List<int> tagIds);
-    Task<Dictionary<int, List<Tag>>> GetTagGroupsByArticleIds(List<int> articleIds);
+    Task<bool> EditArticleTags(long articleId, List<long> tagIds);
+    Task<bool> DeleteByArticleId(long articleId);
+    Task<bool> DeleteByTagId(long tagId);
+    Task<List<ArticleTag>> GetByArticleId(long articleId);
+    Task<List<ArticleTag>> GetByTagId(long tagId);
+    Task<List<ArticleTag>> GetByArticleIds(List<long> articleIds);
+    Task<List<ArticleTag>> GetByTagIds(List<long> tagIds);
+    Task<Dictionary<long, List<Tag>>> GetTagGroupsByArticleIds(List<long> articleIds);
 }
