+++ BackendDotNet/FF.Articles.Backend/FF.Articles.Backend.Contents.API/Models/Dtos/ArticleDto.cs
@@ -3,7 +3,7 @@
 namespace FF.Articles.Backend.Contents.API.Models.Dtos;
 public class ArticleDto
 {
-    public int ArticleId { get; set; }
+    public long ArticleId { get; set; }
     public DateTime? CreateTime { get; set; }
     public DateTime? UpdateTime { get; set; }
 
@@ -14,11 +14,11 @@ public class ArticleDto
     /// 1: article, 2: sub article, 3: topic article
     /// </summary>
     public string ArticleType { get; set; }
-    public int? ParentArticleId { get; set; }
+    public long? ParentArticleId { get; set; }
     public List<ArticleDto> SubArticles { get; set; } = new();
-    public int UserId { get; set; }
+    public long UserId { get; set; }
     public UserApiDto? User { get; set; }
-    public int TopicId { get; set; }
+    public long TopicId { get; set; }
     public string TopicTitle { get; set; } = string.Empty;
     public List<string> Tags { get; set; } = new();
     public int SortNumber { get; set; }
