+++ BackendDotNet/FF.Articles.Backend/FF.Articles.Backend.Contents.API/Models/Dtos/TagDto.cs
@@ -1,7 +1,7 @@
 ﻿namespace FF.Articles.Backend.Contents.API.Models.Dtos;
 public class TagDto
 {
-    public int TagId { get; set; }
+    public long TagId { get; set; }
     public string TagName { get; set; }
 
 }
