+++ BackendDotNet/FF.Articles.Backend/FF.Articles.Backend.Contents.API/Models/Dtos/TopicDto.cs
@@ -3,7 +3,7 @@
 namespace FF.Articles.Backend.Contents.API.Models.Dtos;
 public class TopicDto
 {
-    public int TopicId { get; set; }
+    public long TopicId { get; set; }
     public DateTime? CreateTime { get; set; }
     public DateTime? UpdateTime { get; set; }
 
@@ -12,7 +12,7 @@ public class TopicDto
     public string Content { get; set; } = string.Empty;
     public string Category { get; set; } = string.Empty;
     public string TopicImage { get; set; } = string.Empty;
-    public int UserId { get; set; }
+    public long UserId { get; set; }
     public UserApiDto? User { get; set; }
     public int SortNumber { get; set; }
     public int IsHidden { get; set; }
