+++ BackendDotNet/FF.Articles.Backend/FF.Articles.Backend.Contents.API/Models/Requests/Articles/ArticleAddRequest.cs
@@ -11,8 +11,8 @@ public class ArticleAddRequest
     ///  Article = 1, SubArticle = 2, TopicArticle = 3
     /// </summary>
     public string ArticleType { get; set; } = ArticleTypes.Article;
-    public int? ParentArticleId { get; set; }
-    public int TopicId { get; set; }
+    public long? ParentArticleId { get; set; }
+    public long TopicId { get; set; }
     public List<string> Tags { get; set; } = new();
     public int SortNumber { get; set; } = 1;
     public int IsHidden { get; set; } = 0;
