+++ BackendDotNet/FF.Articles.Backend/FF.Articles.Backend.Contents.API/Models/Requests/Articles/ArticleEditRequest.cs
@@ -3,7 +3,7 @@
 namespace FF.Articles.Backend.Contents.API.Models.Requests.Articles;
 public class ArticleEditRequest
 {
-    public int ArticleId { get; set; }
+    public long ArticleId { get; set; }
     public string? Title { get; set; }
     public string? Content { get; set; }
     public string? Abstract { get; set; }
@@ -11,8 +11,8 @@ public class ArticleEditRequest
     ///  Article = 1, SubArticle = 2, TopicArticle = 3
     /// </summary>
     public string? ArticleType { get; set; }
-    public int? ParentArticleId { get; set; }
-    public int? TopicId { get; set; }
+    public long? ParentArticleId { get; set; }
+    public long? TopicId { get; set; }
     public List<string>? TagNames { get; set; }
     public int? SortNumber { get; set; }
     public int? IsHidden { get; set; }
