+++ BackendDotNet/FF.Articles.Backend/FF.Articles.Backend.Contents.API/Models/Requests/Topics/TopicPageRequest.cs
@@ -4,7 +4,7 @@ namespace FF.Articles.Backend.Contents.API.Models.Requests.Topics
 {
     public class TopicQueryRequest : PageRequest
     {
-        public int? TopicId { get; set; } = null;
+        public long? TopicId { get; set; } = null;
         public bool IncludeUser { get; set; } = false;
         public bool IncludeArticles { get; set; } = false;
         public bool IncludeSubArticles { get; set; } = false;
